Refactor: Create user service, refactor user controller
