+++ server/controllers/user.controller.js
@@ -1,0 +1,11 @@
+const { getUserOffers } = require('../services/user.service')
+
+exports.getUserOffers = async (req, res, next) => {
+  const userId = req.params.id
+  try {
+    const offers = await getUserOffers(userId)
+    res.send(offers)
+  } catch (error) {
+    next(error)
+  }
+}
