+++ server/routes/users.route.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 
-const { getUserOffers } = require('../controllers/users')
+const { getUserOffers } = require('../controllers/user.controller')
 
 router.get('/:id/offers', getUserOffers)
