+++ server/services/user.service.js
@@ -1,0 +1,13 @@
+const Offer = require('../models/Offer')
+
+const UserService = {
+  getUserOffers: async function (userId) {
+    try {
+      return await Offer.find({ creator: userId })
+    } catch (error) {
+      console.log(error)
+    }
+  },
+}
+
+module.exports = UserService
